# alvobot-reference/specs/019-alvoads-meta/contracts/api.yaml
openapi: 3.0.3
info:
  title: AlvoADS Meta API
  description: API endpoints for Meta/Facebook Ads campaign management
  version: 1.0.0
servers:
  - url: /api/meta/campaigns
    description: Meta Campaigns API
paths:
...
  /templates:
    post:
      summary: Save campaign template
      description: Create or update a campaign template (draft)
      operationId: saveTemplate
      tags:
        - Templates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveTemplateRequest'
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignTemplate'
        '200':
          description: Template updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          description: Insufficient credits

    get:
      summary: List campaign templates
      description: Get all campaign templates for the user
      operationId: listTemplates
      tags:
        - Templates
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, ready, publishing, published, failed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of templates
          content:
            application/json:
              schema:
                type: object
                properties:
                  templates:
                    type: array
                    items:
                      $ref: '#/components/schemas/CampaignTemplate'
                  total:
                    type: integer
...
  /{id}/pause:
    post:
      summary: Pause published campaign
      operationId: pauseCampaign
      tags:
        - Publishing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Campaign paused
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    CampaignTemplate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        platform:
          type: string
          enum: [meta_ads]
        name:
          type: string
        status:
          type: string
          enum: [draft, ready, publishing, published, failed]
        campaignData:
          $ref: '#/components/schemas/MetaCampaignData'
        platformIds:
          $ref: '#/components/schemas/MetaPlatformIds'
        creditsConsumed:
          type: integer
        errorMessage:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
    MetaCampaignData:
      type: object
      required:
        - adAccountId
        - pageId
        - name
        - objective
        - targeting
        - budget
        - creative
        - adCopy
      properties:
        adAccountId:
          type: string
        pageId:
          type: string
        instagramAccountId:
          type: string
        name:
          type: string
        objective:
          type: string
          enum: [TRAFFIC, MESSAGES, LEADS, SALES]
        specialAdCategories:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [ACTIVE, PAUSED]
        targeting:
          $ref: '#/components/schemas/MetaTargeting'
        budget:
          $ref: '#/components/schemas/MetaBudget'
        schedule:
          $ref: '#/components/schemas/MetaSchedule'
        creative:
          $ref: '#/components/schemas/MetaCreative'
        adCopy:
          $ref: '#/components/schemas/MetaAdCopy'
        messenger:
          $ref: '#/components/schemas/MetaMessengerConfig'
    MetaTargeting:
      type: object
      required:
        - ageMin
        - ageMax
        - genders
        - countries
        - languages
      properties:
        ageMin:
          type: integer
          minimum: 18
          maximum: 65
        ageMax:
          type: integer
          minimum: 18
          maximum: 65
        genders:
          type: array
          items:
            type: integer
            enum: [0, 1, 2]
        countries:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            $ref: '#/components/schemas/Language'
    MetaBudget:
      type: object
      required:
        - type
        - amount
        - bidStrategy
      properties:
        type:
          type: string
          enum: [daily, lifetime]
        amount:
          type: integer
          minimum: 600
          description: Amount in cents (600 = R$6)
        bidStrategy:
          type: string
          enum: [LOWEST_COST_WITHOUT_CAP, LOWEST_COST_WITH_BID_CAP, COST_CAP]
        costPerResult:
          type: integer
          description: Desired cost per result in cents
    MetaSchedule:
      type: object
      properties:
        type:
          type: string
          enum: [continuous, scheduled]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        timezone:
          type: string
          default: America/Sao_Paulo
    MetaCreative:
      type: object
      required:
        - sourceType
        - images
      properties:
        sourceType:
          type: string
          enum: [google_drive, ai_generated, upload]
        images:
          type: array
          items:
            $ref: '#/components/schemas/CreativeImage'
    CreativeImage:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
        hash:
          type: string
        source:
          type: string
          enum: [google_drive, ai_generated, upload]
        prompt:
          type: string
        width:
          type: integer
        height:
          type: integer
    MetaAdCopy:
      type: object
      required:
        - primaryText
        - headline
        - description
        - callToAction
        - destinationUrl
      properties:
        primaryText:
          type: string
          maxLength: 125
        headline:
          type: string
          maxLength: 27
        description:
          type: string
          maxLength: 27
        callToAction:
          type: string
          enum:
            - LEARN_MORE
            - SHOP_NOW
            - SIGN_UP
            - SUBSCRIBE
            - CONTACT_US
            - GET_OFFER
            - GET_QUOTE
            - BOOK_NOW
            - APPLY_NOW
            - DOWNLOAD
            - SEND_MESSAGE
            - WHATSAPP_MESSAGE
        destinationUrl:
          type: string
          format: uri
        displayUrl:
          type: string
    MetaMessengerConfig:
      type: object
      properties:
        welcomeMessage:
          type: string
        quickReplies:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              title:
                type: string
                maxLength: 20
              payload:
                type: string
    MetaPlatformIds:
      type: object
      properties:
        campaignId:
          type: string
        adSetId:
          type: string
        creativeId:
          type: string
        adId:
          type: string

    # ============================================
    # Data Schemas
    # ============================================
...
    Language:
      type: object
      properties:
        key:
          type: string
        name:
          type: string
...
    SaveTemplateRequest:
      type: object
      required:
        - campaignData
      properties:
        id:
          type: string
          format: uuid
          description: Template ID for updates
        name:
          type: string
        campaignData:
          $ref: '#/components/schemas/MetaCampaignData'
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: string
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...